renamed StructDef to ModelDef and moved it to models.go
moved remaining content in structs.ts to models.ts and deleted file.
generateStruct() was folded into generateStructs().
structDef var names were renamed to modelDef
