--- packages/autorest.go/src/generator/structs.ts
@@ -1,176 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { comment } from '@azure-tools/codegen';
-import { ConstantSchema, Language, SchemaType, Property } from '@autorest/codemodel';
-import { values } from '@azure-tools/linq';
-import { commentLength, isArraySchema } from '../common/helpers';
-import { hasDescription, sortAscending } from './helpers';
-import { ImportManager } from './imports';
-
-// represents a struct method
-export interface StructMethod {
-  name: string;
-  desc: string;
-  text: string;
-}
-
-// represents a struct definition
-export class StructDef {
-  readonly Language: Language;
-  readonly Properties?: Property[];
-  readonly SerDeMethods: StructMethod[];
-  readonly Methods: StructMethod[];
-  HasJSONByteArray: boolean;
-  HasAny: boolean;
-
-  constructor(language: Language, props?: Property[]) {
-    this.Language = language;
-    this.Properties = props;
-    if (this.Properties) {
-      this.Properties.sort((a: Property, b: Property) => { return sortAscending(a.language.go!.name, b.language.go!.name); });
-    }
-    this.SerDeMethods = new Array<StructMethod>();
-    this.Methods = new Array<StructMethod>();
-    this.HasJSONByteArray = false;
-    this.HasAny = false;
-  }
-
-  text(): string {
-    let text = '';
-    if (hasDescription(this.Language)) {
-      text += `${comment(this.Language.description, '// ', undefined, commentLength)}\n`;
-    }
-    text += `type ${this.Language.name} struct {\n`;
-    // used to track when to add an extra \n between fields that have comments
-    let first = true;
-    // group fields by required/optional/read-only in that order
-    this.Properties?.sort((lhs: Property, rhs: Property): number => {
-      if ((lhs.required && !rhs.required) || (!lhs.readOnly && rhs.readOnly)) {
-        return -1;
-      } else if ((rhs.readOnly && !lhs.readOnly) || (!rhs.readOnly && lhs.readOnly)) {
-        return 1;
-      } else {
-        return 0;
-      }
-    });
-    for (const prop of values(this.Properties)) {
-      if (prop.language.go!.embeddedType) {
-        continue;
-      }
-      if (hasDescription(prop.language.go!)) {
-        if (!first) {
-          // add an extra new-line between fields IFF the field
-          // has a comment and it's not the very first one.
-          text += '\n';
-        }
-        text += `\t${comment(prop.language.go!.description, '// ', undefined, commentLength)}\n`;
-      }
-      let typeName = prop.schema.language.go!.name;
-      if (prop.schema.type === SchemaType.Constant) {
-        // for constants we use the underlying type name
-        typeName = (<ConstantSchema>prop.schema).valueType.language.go!.name;
-      }
-      let serialization = prop.serializedName;
-      if (this.Language.marshallingFormat === 'json') {
-        serialization += ',omitempty';
-      } else if (this.Language.marshallingFormat === 'xml') {
-        serialization = getXMLSerialization(prop, this.Language);
-      }
-      let readOnly = '';
-      if (prop.readOnly) {
-        readOnly = ` azure:"ro"`;
-      }
-      let tag = '';
-      // only emit tags for XML; JSON uses custom marshallers/unmarshallers
-      if (this.Language.marshallingFormat === 'xml' && !prop.language.go!.isAdditionalProperties) {
-        tag = ` \`${this.Language.marshallingFormat}:"${serialization}"${readOnly}\``;
-      }
-      text += `\t${prop.language.go!.name} ${getStar(prop.language.go!)}${typeName}${tag}\n`;
-      first = false;
-    }
-    text += '}\n\n';
-    return text;
-  }
-
-  receiverName(): string {
-    const typeName = this.Language.name;
-    return typeName[0].toLowerCase();
-  }
-}
-
-export function generateStruct(imports: ImportManager, lang: Language, props?: Property[]): StructDef {
-  if (lang.isLRO) {
-    imports.add('time');
-    imports.add('context');
-  }
-  const st = new StructDef(lang, props);
-  for (const prop of values(props)) {
-    imports.addImportForSchemaType(prop.schema);
-    if (prop.schema.type === SchemaType.Any && !prop.schema.language.go!.rawJSONAsBytes) {
-      st.HasAny = true;
-    }
-  }
-  return st;
-}
-
-export function getStar(lang: Language): string {
-  // lang is assumed to be go
-  if (lang.byValue === true) {
-    return '';
-  }
-  return '*';
-}
-
-export function getXMLSerialization(prop: Property, lang: Language): string {
-  let serialization = prop.serializedName;
-  // default to using the serialization name
-  if (prop.schema.serialization?.xml?.name) {
-    // xml can specifiy its own name, prefer that if available
-    serialization = prop.schema.serialization.xml.name;
-  } else if (prop.schema.serialization?.xml?.text) {
-    // type has the x-ms-text attribute applied so it should be character data, not a node (https://github.com/Azure/autorest/tree/main/docs/extensions#x-ms-text)
-    // see https://pkg.go.dev/encoding/xml#Unmarshal for what ,chardata actually means
-    serialization = ',chardata';
-  }
-  if (prop.schema.serialization?.xml?.attribute) {
-    // value comes from an xml attribute
-    serialization += ',attr';
-  } else if (isArraySchema(prop.schema)) {
-    // start with the serialized name of the element, preferring xml name if available
-    let inner = prop.schema.elementType.language.go!.name;
-    if (prop.schema.elementType.serialization?.xml?.name) {
-      inner = prop.schema.elementType.serialization.xml.name;
-    }
-    // arrays can be wrapped or unwrapped.  here's a wrapped example
-    // note how the array of apple objects is "wrapped" in GoodApples
-    // <AppleBarrel>
-    //   <GoodApples>
-    //     <Apple>Fuji</Apple>
-    //     <Apple>Gala</Apple>
-    //   </GoodApples>
-    // </AppleBarrel>
-
-    // here's an unwrapped example, the array of slide objects
-    // is embedded directly in the object (no "wrapping")
-    // <slideshow>
-    //   <slide>
-    //     <title>Wake up to WonderWidgets!</title>
-    //   </slide>
-    //   <slide>
-    //     <title>Overview</title>
-    //   </slide>
-    // </slideshow>
-
-    // arrays in the response type are handled slightly different as we
-    // unmarshal directly into them so no need to add the unwrapping.
-    if (prop.schema.serialization?.xml?.wrapped && lang.responseType !== true) {
-      serialization += `>${inner}`;
-    } else {
-      serialization = inner;
-    }
-  }
-  return serialization;
-}
